server/requirements.txt
Flask
flask-mongoengine
dotenv
datetime
flask-cors